+++ src/Symfony/Bundle/SecurityBundle/Command/UserPasswordEncoderCommand.php
@@ -73,7 +73,7 @@ Suppose that you have the following security configuration in your application:
 # app/config/security.yml
 security:
     encoders:
-        Symfony\Component\Security\Core\User\User: plaintext
+        Symfony\Component\Security\Core\User\InMemoryUser: plaintext
         App\Entity\User: auto
 </comment>