+++ src/Symfony/Component/Security/Core/Tests/User/UserTest.php
@@ -16,6 +16,9 @@ use Symfony\Component\Security\Core\User\EquatableInterface;
 use Symfony\Component\Security\Core\User\User;
 use Symfony\Component\Security\Core\User\UserInterface;
 
+/**
+ * @group legacy
+ */
 class UserTest extends TestCase
 {
     public function testConstructorException()