+++ src/Symfony/Component/Security/Guard/Tests/Authenticator/GuardBridgeAuthenticatorTest.php
@@ -16,7 +16,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
-use Symfony\Component\Security\Core\User\User;
+use Symfony\Component\Security\Core\User\InMemoryUser;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
 use Symfony\Component\Security\Guard\Authenticator\GuardBridgeAuthenticator;
 use Symfony\Component\Security\Guard\AuthenticatorInterface;
@@ -77,7 +77,7 @@ class GuardBridgeAuthenticatorTest extends TestCase
             ->with($request)
             ->willReturn($credentials);
 
-        $user = new User('test', null, ['ROLE_USER']);
+        $user = new InMemoryUser('test', null, ['ROLE_USER']);
         $this->guardAuthenticator->expects($this->once())
             ->method('getUser')
             ->with($credentials, $this->userProvider)
@@ -145,7 +145,7 @@ class GuardBridgeAuthenticatorTest extends TestCase
 
     public function testCreateAuthenticatedToken()
     {
-        $user = new User('test', null, ['ROLE_USER']);
+        $user = new InMemoryUser('test', null, ['ROLE_USER']);
 
         $token = new PostAuthenticationGuardToken($user, 'main', ['ROLE_USER']);
         $this->guardAuthenticator->expects($this->once())
@@ -159,7 +159,7 @@ class GuardBridgeAuthenticatorTest extends TestCase
     public function testHandleSuccess()
     {
         $request = new Request();
-        $token = new PostAuthenticationGuardToken(new User('test', null, ['ROLE_USER']), 'main', ['ROLE_USER']);
+        $token = new PostAuthenticationGuardToken(new InMemoryUser('test', null, ['ROLE_USER']), 'main', ['ROLE_USER']);
 
         $response = new Response();
         $this->guardAuthenticator->expects($this->once())